Json formatting mostly done just need to write up the context ints

+++ src/models.rs
@@ -14,13 +14,20 @@ pub struct TokenJson {
     pub created_at: String,
     pub response: String,
     pub done: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub context: Option<Vec<i32>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub total_duration: Option<i64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
     pub load_duration: Option<i64>,
-    pub prompt_eval_count: Option<i32>,
-    pub prompt_eval_duration: Option<i64>,
-    pub eval_count: Option<i32>,
-    pub eval_duration: Option<i64>
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub prompt_eval_count: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub prompt_eval_duration: Option<u128>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub eval_count: Option<usize>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub eval_duration: Option<u128>
 }
 
 #[derive(Serialize, Deserialize)]
